third_party/blink/web_tests/platform/win11-arm64/http/tests/security/script-crossorigin-loads-cross-origin-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
Test that a script element with a crossorigin attribute loads a cross-origin script correctly when the CORS check passes on the response.

FAIL
